app/requirements.txt
fastapi[standard]
uvicorn
pydantic
pandas
scikit-learn
joblib
python-dotenv